move readEditorTabStyle() to IntUiBridgeTab
