+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridgeTab.kt
@@ -69,3 +69,56 @@ internal fun readDefaultTabStyle(): TabStyle {
         scrollbarStyle = readScrollbarStyle(isDark),
     )
 }
+
+internal fun readEditorTabStyle(): TabStyle {
+    val normalBackground = JBUI.CurrentTheme.EditorTabs.background().toComposeColor()
+    val selectedBackground = JBUI.CurrentTheme.EditorTabs.underlinedTabBackground().toComposeColorOrUnspecified()
+    val normalContent = retrieveColorOrUnspecified("TabbedPane.foreground")
+    val selectedUnderline = retrieveColorOrUnspecified("TabbedPane.underlineColor")
+
+    val colors =
+        TabColors(
+            background = normalBackground,
+            backgroundDisabled = normalBackground,
+            backgroundPressed = selectedBackground,
+            backgroundHovered = JBUI.CurrentTheme.EditorTabs.hoverBackground().toComposeColor(),
+            backgroundSelected = selectedBackground,
+            content = normalContent,
+            contentDisabled = retrieveColorOrUnspecified("TabbedPane.disabledForeground"),
+            contentPressed = normalContent,
+            contentHovered = normalContent,
+            contentSelected = normalContent,
+            underline = Color.Unspecified,
+            underlineDisabled = retrieveColorOrUnspecified("TabbedPane.disabledUnderlineColor"),
+            underlinePressed = selectedUnderline,
+            underlineHovered = Color.Unspecified,
+            underlineSelected = selectedUnderline,
+        )
+
+    return TabStyle(
+        colors = colors,
+        metrics =
+            TabMetrics(
+                underlineThickness = retrieveIntAsDpOrUnspecified("TabbedPane.tabSelectionHeight").takeOrElse { 2.dp },
+                tabPadding = retrieveInsetsAsPaddingValues("TabbedPane.tabInsets"),
+                closeContentGap = 4.dp,
+                tabContentSpacing = 4.dp,
+                tabHeight = retrieveIntAsDpOrUnspecified("TabbedPane.tabHeight").takeOrElse { 24.dp },
+            ),
+        icons = TabIcons(close = AllIconsKeys.General.CloseSmall),
+        contentAlpha =
+            TabContentAlpha(
+                iconNormal = .7f,
+                iconDisabled = .7f,
+                iconPressed = 1f,
+                iconHovered = 1f,
+                iconSelected = 1f,
+                contentNormal = .7f,
+                contentDisabled = .7f,
+                contentPressed = 1f,
+                contentHovered = 1f,
+                contentSelected = 1f,
+            ),
+        scrollbarStyle = readScrollbarStyle(isDark),
+    )
+}
